net/ntnic: improve logging format specifiers
Updated logging format specifiers for better consistency
and to match the expected data types

+++ drivers/net/ntnic/ntnic_filter/ntnic_filter.c
@@ -355,8 +355,8 @@ int nthw_create_action_elements_inline(struct cnv_action_s *action,
 				case RTE_ETH_HASH_FUNCTION_MAX:
 				default:
 					NT_LOG(ERR, FILTER,
-						"RTE ACTION RSS - unsupported function: %u",
-						rss->func);
+						"RTE ACTION RSS - unsupported function: %i",
+						(int)rss->func);
 					return -1;
 				}
